Add inline rating to recipes in planner view
When viewing a day's planned meals, allow the user to rate the recipe
via clicking the star. This is initially using a link to a redirect
view that handles updating the data, but it will eventually be
done via javascript, so as to not interrupt the user's experience
with refreshes.